Progress: Complete `2-mmio-trap` Complete the `CA2025: Exercise` tasks in the `2-mmio-trap lab`. All tests pass using make test and make compliance, which correspond to the Chisel unit tests and RISCOF compliance tests, respectively.
Most changes are similar to those in `1-single-cycle`, except
that this lab requires the use of the CSR extension and machine-
mode implementation. In addition, VGA display support is added,
and also the requirement of updating the `CROSS_COMPILE`
parameter in theMakefile to be able rebuild the nyancat animation
data.